requirements.txt
# DevMesh Platform - Python Dependencies

# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
pymysql==1.1.0
aiomysql==0.2.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
requests==2.31.0
pyyaml>=6.0

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
